feat: add api to get podcast information
The class Provider (API pkg) is in charge of getting all the info
needed related to the Podcast (podcast info, episodes, patreon..).

The Model pkg contains the different classes the represent the info
obtained by Provider.

This commit adds the changes described above along with several
files as this is the first application commit.