feat: add ViewModel
The ViewModel class provides LiveData objects for the different
information that will be presented to the user in the screen.

The UI will be performed in next commits by using the Navigation
Component.

The PatreonTier model, and getPatreonInfo method in Provider.kt have
been modified due to changes in the PyC Patreon page.

This commit add the changes and new files needed for the work described
above.