feat: start UI work
This commit adds the following:

* UI initial state
Currently the UI is composed by two fragments (home and upcoming).
The navigation between the fragments is preformed by using Navigation
Component and a bottom navigation bar.

Is possible to share the upcoming information as an image. This part
needs to be reviewed in future commits.

Different UI resources (layouts, drawing, menus) have been added from
the PyC java repository.

* README update
The readme file is updated with information about the library and
external resources used.