feat: add SocialNetwork and Patreon fragments
This commit adds the following:

* SocialNetwork fragment
This fragment displays the different social networks when you can find
Planos y Centellas.

The user is able to click on each icon, opening the corresponding app
or web browser.

* Patreon fragment
This fragment contains the 3 different Patreon options available along
with a button to subscribe to each one.

The subscription process is performed in the Patreon web site.

* README update
The readme file is updated with new external resources.