feat: add EpisodeListFragment
This commit adds the following:

* EpisodeList fragment
This fragment shows all the Planos y Centellas' episodes published in
Ivoox.

The episodes are set in the screen by using a RecyclerView.

Each episode contain its name and image. If the user click on one of
them, the Ivoox application will be opened to listener the desired
episode.

Is possible to search for an specific episode by using the search icon
in the top bar. The search feature needs to improve (remove accent
mark of the episode names and query)

* Fix patreon fragment bug
The title of the patreon tier wasn't set correctly. This has been fixed.

* Model updates

Some models have been updated.