feat: fix memory leaks and add longClickListener
The memory leak was related to databinding being not released when a
fragment was destroyed/we moved to a different one.

To fix it, I followed this link:
https://proandroiddev.com/avoiding-memory-leaks-when-using-data-binding
-and-view-binding-3b91d571c150

Besides this, a longClick listener has been installed in the
EpisodeListFragment.kt to share the selected episode with friends.

Finally, due to issues with Picasso loading some images,
usesCleartextTraffic has been set to true in the Manifest file. This will
change in future commits.